+++ src/main/java/raisetech/student/management/Application.java
@@ -1,40 +1,17 @@
 package raisetech.student.management;
 
-import java.util.List;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+
 
 @SpringBootApplication
 
-@RestController
 public class Application {
 
-	@Autowired
-	private StudentRepository repository;
-
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-	@GetMapping("studentList")
-	public List<Student> getStudentList(){
-		 return repository.search();
-
-	}
-
-	@GetMapping("studentCoursesList")
-	public List<StudentCourses> getStudentCoursesList() {
-		return repository.searchCourses();
-	}
-
-
-
-
+  public static void main(String[] args) {
+    SpringApplication.run(Application.class, args);
+  }
 
 
 }
+
